First draft of orders index view
Created a first draft of the orders index view and wired it to serve at
/orders. To accomplish this we created an OrdersContreller with an index
action.

Other thoughts: The full creating an order feature (issue #13) can't be
implemented at this point because we don't have a users table
(@mdorrance and @imwithsam are working on authentication) and to create
the the orders table we need to be able to create it with a foreign key
(user_id) which we can't do without a users table.