Use #cart method vs @cart in cart_items index
https://waffle.io/michael-reeves/redrum_nursery/cards/55c11be36134550f00ccd575

Why:
As we were working on issue #13 (Creating an order) we found that we
were using @cart in the cart_items index view vs #cart (calling the
method). This made it so that when stubbing out #cart in the middle of our
user_creates_an_order_spec, the cart_items index view page wouldn't
render correctly as it returned nil for @cart.

This change addresses the need by:

* making #cart a helper method in application_controller
* changing @cart for #cart in cart_items index view