Add admin view of order details
Why:
Admin users can see the order's date and time.
And can see the purchaser's full name and address.
And can see, for each product on the order:
- The product's name, which is linked to the product page.
- Quantity in this order.
- Price
- Line item subtotal.
And can see the total for the order.
And can see the status for the order.

This change addresses the need by:
* add routes for admin/orders #index and #show actions
* add the #show action to Admin::OrdersController
* refactor orders/show page to move the table to a partial
* add address partial
* add admin/orders/show page to include single_order_table partial
* add associated spec for viewing the Admin::Orders #show page

: