Add submit payment feature(stripe) upon checkout
https://waffle.io/michael-reeves/redrum_nursery/cards/55c7f0773c538f320083ffca

Why:
Per user story #52, a user should submit a payment after clicking
checkout while in the cart index view.

This change addresses the need by:
* adding ChargesController with a #create action that charges customer and triggers the
  creation of an order and order_items if the charge is successful
* adding coniditional to cart index view that displays checkout button
  only for logged in users and users who have products in the cart
* adding checkout custom stripe checkout button on cart index view
* removing #create action in OrdersController as it's not needed anymore
* removing create resource path in routes (not needed anymore)
* modified existing tests for them to test new functionality

Important notes: The user_creates_an_orders_spec test, that checks if a
user can place an order, was skipped as the team is not clear if we
should test the stripe integration. We're afraid of being shut down by
Stripe if we call their API too often.